+++ common/src/main/java/earth/terrarium/cadmus/common/util/ModUtils.java
@@ -1,5 +1,6 @@
 package earth.terrarium.cadmus.common.util;
 
+import com.teamresourceful.resourcefullib.common.utils.CommonUtils;
 import earth.terrarium.cadmus.common.claims.AdminClaimHandler;
 import earth.terrarium.cadmus.common.claims.ClaimHandler;
 import earth.terrarium.cadmus.common.claims.admin.ModFlags;
@@ -7,6 +8,7 @@ import earth.terrarium.cadmus.common.constants.ConstantComponents;
 import earth.terrarium.cadmus.common.teams.TeamHelper;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
@@ -16,6 +18,7 @@ import net.minecraft.world.level.ChunkPos;
 import java.util.Objects;
 
 public class ModUtils {
+
     public static void displayTeamName(ServerPlayer player, ChunkPos pos) {
         if (!(player instanceof LastMessageHolder holder)) return;
 
@@ -52,6 +55,20 @@ public class ModUtils {
 
     public static GameProfileCache getProfileCache(MinecraftServer server) {
         return Objects.requireNonNull(server.getProfileCache());
+    }
+
+    public static Component translatableWithStyle(String key, Object... args) {
+        for (int i = 0; i < args.length; ++i) {
+            if (!(args[i] instanceof MutableComponent component)) continue;
+            if (component.getStyle().getColor() == null) continue;
+
+            ChatFormatting color = ChatFormatting.getByName(component.getStyle().getColor().toString());
+            if (color != null) {
+                args[i] = "§" + color.getChar() + component.getString();
+            }
+        }
+
+        return Component.literal(CommonUtils.serverTranslatable(key, args).getString());
     }
 
     public static boolean isAdmin(String id) {